Move DB and related methods from ContainerSchemaManager to ReconContainerManager.
