+++ hadoop-ozone/recon/src/main/java/org/apache/hadoop/ozone/recon/scm/ReconContainerManager.java
@@ -18,12 +18,17 @@
 
 package org.apache.hadoop.ozone.recon.scm;
 
+import static java.util.Comparator.comparingLong;
 import static org.apache.hadoop.hdds.protocol.proto.HddsProtos.LifeCycleEvent.FINALIZE;
 
 import java.io.IOException;
+import java.util.ArrayList;
+import java.util.HashMap;
+import java.util.List;
 import java.util.Map;
 import java.util.UUID;
 import java.util.concurrent.ConcurrentHashMap;
+import java.util.stream.Collectors;
 
 import com.google.common.annotations.VisibleForTesting;
 import org.apache.hadoop.hdds.conf.ConfigurationSource;
@@ -41,11 +46,15 @@ import org.apache.hadoop.hdds.scm.pipeline.PipelineID;
 import org.apache.hadoop.hdds.scm.pipeline.PipelineManager;
 import org.apache.hadoop.hdds.utils.db.DBStore;
 import org.apache.hadoop.hdds.utils.db.Table;
+import org.apache.hadoop.ozone.recon.persistence.ContainerHistory;
 import org.apache.hadoop.ozone.recon.persistence.ContainerSchemaManager;
+import org.apache.hadoop.ozone.recon.spi.ContainerDBServiceProvider;
 import org.apache.hadoop.ozone.recon.spi.StorageContainerServiceProvider;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+
+import javax.inject.Inject;
 
 /**
  * Recon's overriding implementation of SCM's Container Manager.
@@ -54,9 +63,13 @@ public class ReconContainerManager extends SCMContainerManager {
 
   private static final Logger LOG =
       LoggerFactory.getLogger(ReconContainerManager.class);
-  private StorageContainerServiceProvider scmClient;
-  private ContainerSchemaManager containerSchemaManager;
+  private final StorageContainerServiceProvider scmClient;
+  private final ContainerSchemaManager containerSchemaManager;
 
+  @Inject
+  private ContainerDBServiceProvider cdbServiceProvider;
+
+  private final Table<UUID, DatanodeDetails> nodeDB;
   // Container ID -> Datanode UUID -> Timestamp
   private final Map<Long, Map<UUID, ContainerReplicaHistory>> replicaHistoryMap;
 
@@ -81,8 +94,8 @@ public class ReconContainerManager extends SCMContainerManager {
     this.scmClient = scm;
     this.containerSchemaManager = containerSchemaManager;
     this.replicaHistoryMap = new ConcurrentHashMap<>();
-    containerSchemaManager.setReplicaHistoryMap(replicaHistoryMap);
-    containerSchemaManager.setScmDBStore(batchHandler);
+    // batchHandler == scmDBStore
+    this.nodeDB = ReconSCMDBDefinition.NODES.getTable(batchHandler);
   }
 
   /**
@@ -223,7 +236,7 @@ public class ReconContainerManager extends SCMContainerManager {
     }
 
     if (flushToDB) {
-      containerSchemaManager.upsertContainerHistory(id, uuid, currTime);
+      upsertContainerHistory(id, uuid, currTime);
     }
   }
 
@@ -246,8 +259,7 @@ public class ReconContainerManager extends SCMContainerManager {
       final ContainerReplicaHistory ts = replicaLastSeenMap.get(uuid);
       if (ts != null) {
         // Flush to DB, then remove from in-memory map
-        containerSchemaManager.upsertContainerHistory(id, uuid,
-            ts.getLastSeenTime());
+        upsertContainerHistory(id, uuid, ts.getLastSeenTime());
         replicaLastSeenMap.remove(uuid);
       }
     }
@@ -263,12 +275,102 @@ public class ReconContainerManager extends SCMContainerManager {
     return replicaHistoryMap;
   }
 
+  public List<ContainerHistory> getAllContainerHistory(long containerID) {
+    // First, get the existing entries from DB
+    Map<UUID, ContainerReplicaHistory> resMap;
+    try {
+      resMap = cdbServiceProvider.getContainerReplicaHistory(containerID);
+    } catch (IOException ex) {
+      resMap = new HashMap<>();
+      LOG.debug("Unable to retrieve container replica history from RDB.");
+    }
+
+    // Then, update the entries with the latest in-memory info, if available
+    if (replicaHistoryMap != null) {
+      Map<UUID, ContainerReplicaHistory> replicaLastSeenMap =
+          replicaHistoryMap.get(containerID);
+      if (replicaLastSeenMap != null) {
+        Map<UUID, ContainerReplicaHistory> finalResMap = resMap;
+        replicaLastSeenMap.forEach((k, v) ->
+            finalResMap.merge(k, v, (old, latest) -> latest));
+        resMap = finalResMap;
+      }
+    }
+
+    // Finally, convert map to list for output
+    List<ContainerHistory> resList = new ArrayList<>();
+    for (Map.Entry<UUID, ContainerReplicaHistory> entry : resMap.entrySet()) {
+      final UUID uuid = entry.getKey();
+      String hostname = "N/A";
+      // Attempt to retrieve hostname from NODES table
+      if (nodeDB != null) {
+        try {
+          DatanodeDetails dnDetails = nodeDB.get(uuid);
+          if (dnDetails != null) {
+            hostname = dnDetails.getHostName();
+          }
+        } catch (IOException ex) {
+          LOG.debug("Unable to retrieve from NODES table of node {}. {}",
+              uuid, ex.getMessage());
+        }
+      }
+      final long firstSeenTime = entry.getValue().getFirstSeenTime();
+      final long lastSeenTime = entry.getValue().getLastSeenTime();
+      resList.add(new ContainerHistory(containerID, uuid.toString(), hostname,
+          firstSeenTime, lastSeenTime));
+    }
+    return resList;
+  }
+
+  public List<ContainerHistory> getLatestContainerHistory(long containerID,
+      int limit) {
+    List<ContainerHistory> res = getAllContainerHistory(containerID);
+    res.sort(comparingLong(ContainerHistory::getLastSeenTime).reversed());
+    return res.stream().limit(limit).collect(Collectors.toList());
+  }
+
   /**
    * Flush the container replica history in-memory map to DB.
    * Expected to be called on Recon graceful shutdown.
    * @param clearMap true to clear the in-memory map after flushing completes.
    */
   public void flushReplicaHistoryMapToDB(boolean clearMap) {
-    containerSchemaManager.flushReplicaHistoryMapToDB(clearMap);
+    if (replicaHistoryMap == null) {
+      return;
+    }
+    synchronized (replicaHistoryMap) {
+      try {
+        cdbServiceProvider.batchStoreContainerReplicaHistory(replicaHistoryMap);
+      } catch (IOException e) {
+        LOG.debug("Error flushing container replica history to DB. {}",
+            e.getMessage());
+      }
+      if (clearMap) {
+        replicaHistoryMap.clear();
+      }
+    }
   }
+
+  public void upsertContainerHistory(long containerID, UUID uuid, long time) {
+    Map<UUID, ContainerReplicaHistory> tsMap;
+    try {
+      tsMap = cdbServiceProvider.getContainerReplicaHistory(containerID);
+      ContainerReplicaHistory ts = tsMap.get(uuid);
+      if (ts == null) {
+        // New entry
+        tsMap.put(uuid, new ContainerReplicaHistory(uuid, time, time));
+      } else {
+        // Entry exists, update last seen time and put it back to DB.
+        ts.setLastSeenTime(time);
+      }
+      cdbServiceProvider.storeContainerReplicaHistory(containerID, tsMap);
+    } catch (IOException e) {
+      LOG.debug("Error on DB operations. {}", e.getMessage());
+    }
+  }
+
+  public Table<UUID, DatanodeDetails> getNodeDB() {
+    return nodeDB;
+  }
+
 }
