+++ hadoop-ozone/recon/src/test/java/org/apache/hadoop/ozone/recon/fsck/TestContainerHealthTask.java
@@ -69,8 +69,7 @@ public class TestContainerHealthTask extends AbstractReconSqlDBTest {
     ContainerSchemaManager containerSchemaManager =
         new ContainerSchemaManager(
             getSchemaDefinition(ContainerSchemaDefinition.class),
-            unHealthyContainersTableHandle,
-            mock(ContainerDBServiceProvider.class));
+            unHealthyContainersTableHandle);
     ReconStorageContainerManagerFacade scmMock =
         mock(ReconStorageContainerManagerFacade.class);
     MockPlacementPolicy placementMock = new MockPlacementPolicy();
